+++ lib/ash_workflow/scheduler/precise/timeline.ex
@@ -455,4 +455,5 @@ defmodule AshWorkflow.Scheduler.Precise.Timeline do
   defp singular(:hours), do: :hour
   defp singular(:minutes), do: :minute
   defp singular(:seconds), do: :second
+  defp singular(:milliseconds), do: :millisecond
 end